fixed varietyjs installation bug

+++ src/install/common.py
@@ -43,7 +43,7 @@ def main(distribution):  # pylint: disable=too-many-statements
     if not distribution == 'xenial':
         pip3_install_packages('testresources')
 
-    # get a bugfree recent pip version
+    # get a bug free recent pip version
     apt_remove_packages('python3-pip', 'python3-setuptools', 'python3-wheel')
     install_pip('python3')
 
@@ -52,7 +52,7 @@ def main(distribution):  # pylint: disable=too-many-statements
     apt_remove_packages('python-pip')
     install_pip('python2')
 
-    # install general python dependencys
+    # install general python dependencies
     apt_install_packages('libmagic-dev')
     apt_install_packages('libfuzzy-dev')
     apt_install_packages('python3-tlsh')
@@ -65,16 +65,14 @@ def main(distribution):  # pylint: disable=too-many-statements
 
     pip3_install_packages('requests')
 
-    # install python mongo bindings
+    # install python MongoDB bindings
     pip3_install_packages('pymongo', 'pyyaml')
 
     # VarietyJS (is executed by update_statistic.py)
-    try:
-        install_github_project('variety/variety', ['git checkout 2f4d815', 'mv -f variety.js ../../bin', 'mv -f spec ../../bin'])
-    except InstallationError as installation_error:
-        if 'Directory not empty' not in str(installation_error):
-            raise installation_error
+    if Path('../bin/spec').exists():
         logging.warning('variety spec not overwritten')
+    else:
+        install_github_project('variety/variety', ['git checkout 2f4d815', 'mv -f variety.js ../../bin', 'mv -f spec ../../bin'])
 
     #  installing common code modules
     pip3_install_packages('hurry.filesize')
